+++ src/SendHeaders.php
@@ -40,7 +40,7 @@ class SendHeaders implements \Dxw\Iguana\Registerable
 		if (count($this->pageProperties)) {
 			// if we are logged in, or on the front page we don't need to worry about configuring things further
 			if (
-				($this->pageProperties['isLoggedInUser'] || $this->pageProperties['requiresPassword'])
+				($this->pageProperties['isLoggedInUser'] || $this->pageProperties['requiresPassword'] || $this->pageProperties['isPreviewPage'])
 				&& !$this->developerMode
 			) {
 				header('Cache-Control: no-cache, no-store, private');
@@ -124,6 +124,7 @@ class SendHeaders implements \Dxw\Iguana\Registerable
 			'isFrontPage' => is_front_page(),
 			'isHomePage' => is_home(),
 			'isLoggedInUser' => is_user_logged_in(),
+			'isPreviewPage' => is_preview(),
 			'postType' => get_post_type() ?? 'unknown',
 			'taxonomies' => get_post_taxonomies() ?? ['none'],
 			'templateName' => get_page_template_slug() ?: 'default',