+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Lowercased company name in plugin metadata
+- Bug where preview pages had public cache headers set
 
 ## [v1.0.0] - 2024-12-10